+++ Source/Core/Visualization/Object/ObjectBase.h
@@ -39,30 +39,31 @@ public:
     };
 
 private:
-    ObjectType m_object_type; ///< object type
-    std::string m_name; ///< object name
-    kvs::Vec3 m_min_object_coord; ///< min coord in the object coordinate system
-    kvs::Vec3 m_max_object_coord; ///< max coord in the object coordinate system
-    kvs::Vec3 m_min_external_coord; ///< min coord in the external coordinate system
-    kvs::Vec3 m_max_external_coord; ///< max coord in the external coordinate system
-    bool m_has_min_max_object_coords; ///< has min-max coorinate values ?
-    bool m_has_min_max_external_coords; ///< has min-max coorinate values ?
-    kvs::Vec3 m_object_center; ///< center of the object in object coordinate system
-    kvs::Vec3 m_external_center; ///< center of the object in external object coordinate system
-    kvs::Vec3 m_normalize; ///< normalize parameter
-    bool m_show_flag; ///< flag for showing object
+    ObjectType m_object_type = UnknownObject; ///< object type
+    std::string m_name = "unknown"; ///< object name
+    kvs::Vec3 m_min_object_coord{ -3.0f, -3.0f, -3.0f }; ///< min coord in the object coordinate system
+    kvs::Vec3 m_max_object_coord{  3.0f,  3.0f,  3.0f }; ///< max coord in the object coordinate system
+    kvs::Vec3 m_min_external_coord{ -3.0f, -3.0f, -3.0f }; ///< min coord in the external coordinate system
+    kvs::Vec3 m_max_external_coord{  3.0f,  3.0f,  3.0f }; ///< max coord in the external coordinate system
+    bool m_has_min_max_object_coords = false; ///< has min-max coorinate values ?
+    bool m_has_min_max_external_coords = false; ///< has min-max coorinate values ?
+    kvs::Vec3 m_object_center{ 0.0f, 0.0f, 0.0f }; ///< center of the object in object coordinate system
+    kvs::Vec3 m_external_center{ 0.0f, 0.0f, 0.0f }; ///< center of the object in external object coordinate system
+    kvs::Vec3 m_normalize{ 1.0f, 1.0f, 1.0f }; ///< normalize parameter
+    bool m_visible = true; ///< visiblity for the object
 
 public:
-    ObjectBase();
-    virtual ~ObjectBase();
+    ObjectBase() = default;
+    virtual ~ObjectBase() = default;
 
     ObjectBase& operator = ( const ObjectBase& object );
 
     void setName( const std::string& name ) { m_name = name; }
     void setMinMaxObjectCoords( const kvs::Vec3& min_coord, const kvs::Vec3& max_coord );
     void setMinMaxExternalCoords( const kvs::Vec3& min_coord, const kvs::Vec3& max_coord );
-    void show() { m_show_flag = true; }
-    void hide() { m_show_flag = false; }
+    void setVisible( const bool visible = true ) { m_visible = visible; }
+    void show() { this->setVisible( true ); }
+    void hide() { this->setVisible( false ); }
     virtual void print( std::ostream& os, const kvs::Indent& indent = kvs::Indent(0) ) const;
     virtual bool read( const std::string& filename );
     virtual bool write( const std::string& filename, const bool ascii = true, const bool external = false ) const;
@@ -79,7 +80,8 @@ public:
     const kvs::Vec3& externalCenter() const { return m_external_center; }
     const kvs::Vec3& normalize() const { return m_normalize; }
     const kvs::Mat4 modelingMatrix() const { return this->xform().toMatrix(); }
-    bool isShown() const { return m_show_flag; }
+    bool isVisible() const { return m_visible; }
+//    bool isShown() const { return m_visible; }
 
     void updateNormalizeParameters();
     virtual void updateMinMaxCoords() {};
@@ -90,7 +92,6 @@ protected:
     void setNormalize( const kvs::Vec3& normalize ) { m_normalize = normalize; }
 
 public:
-
     KVS_DEPRECATED( ObjectBase( const kvs::Vec3& translation, const kvs::Vec3& scaling, const kvs::Mat3& rotation ) )
     {
         m_name = std::string("unknown");
@@ -100,7 +101,7 @@ public:
         m_max_external_coord = kvs::Vec3(  3.0,  3.0,  3.0 );
         m_has_min_max_object_coords = false;
         m_has_min_max_external_coords = false;
-        m_show_flag = true;
+        m_visible = true;
 
         this->setXform( kvs::Xform( translation, scaling, rotation ) );
         this->saveXform();
@@ -123,8 +124,8 @@ public:
     }
 
     KVS_DEPRECATED( friend std::ostream& operator << ( std::ostream& os, const ObjectBase& object ) );
-
     KVS_DEPRECATED( const kvs::Vec3& externalPosition() const ) { return this->externalCenter(); }
+    KVS_DEPRECATED( bool isShown() const ) { return this->isVisible(); }
 };
 
 } // end of namespace kvs